admission: move sqlNodeCPUOverloadIndicator into its own file
Pure code movement.

Release note: None
